# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DomusUnify API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/lists": {
      "get": {
        "tags": [
          "Lists"
        ],
        "summary": "Obtém todas as listas visíveis para o utilizador na família atual.",
        "description": "Lista todas as listas visíveis para o utilizador, incluindo contagem de itens, concluídos e informação de partilha.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lists"
        ],
        "summary": "Cria uma nova lista para a família atual.",
        "description": "Permite definir nome, cor, tipo e visibilidade da lista.",
        "requestBody": {
          "description": "O pedido contendo detalhes da lista a criar.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lists/{listId}": {
      "patch": {
        "tags": [
          "Lists"
        ],
        "summary": "Atualiza uma lista existente.",
        "description": "Permite alterar nome, cor e tipo da lista. A visibilidade/partilha pode ser alterada (opcional).",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "O ID da lista a atualizar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes atualizados da lista.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Lists"
        ],
        "summary": "Elimina uma lista existente.",
        "description": "Ao eliminar uma lista, são eliminados também os respetivos itens.",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "O ID da lista a eliminar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/lists/{listId}/items": {
      "get": {
        "tags": [
          "Lists"
        ],
        "summary": "Obtém todos os itens de uma lista.",
        "description": "Lista todos os itens da lista especificada, incluindo estado de conclusão e categoria.",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "O ID da lista.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListItemResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListItemResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListItemResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lists"
        ],
        "summary": "Adiciona um novo item a uma lista.",
        "description": "Permite definir nome e categoria opcional do item.",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "O ID da lista.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes do item a adicionar.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lists/items/{itemId}": {
      "patch": {
        "tags": [
          "Lists"
        ],
        "summary": "Atualiza um item de lista existente.",
        "description": "Permite alterar nome, estado de conclusão e categoria. CategoryId pode ser null para remover categoria.",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "O ID do item a atualizar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes atualizados do item.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Lists"
        ],
        "summary": "Elimina um item de lista existente.",
        "description": "Remove o item da lista.",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "O ID do item a eliminar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateListItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do item.",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "description": "Identificador da categoria do item (opcional).",
            "format": "uuid",
            "nullable": true
          },
          "assigneeUserId": {
            "type": "string",
            "description": "Identificador do utilizador a quem o item estÃ¡ atribuÃ­do (opcional).",
            "format": "uuid",
            "nullable": true
          },
          "note": {
            "type": "string",
            "description": "Nota opcional associada ao item.",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "description": "URL (ou data URL) de uma foto associada ao item (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para criar um novo item numa lista."
      },
      "CreateListRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da lista.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Tipo de lista (`Shopping`, `Tasks` ou `Custom`).",
            "nullable": true
          },
          "colorHex": {
            "type": "string",
            "description": "Cor da lista em hexadecimal (ex.: `#FFAA00`).",
            "nullable": true
          },
          "visibilityMode": {
            "type": "string",
            "description": "Modo de visibilidade da lista (`Private`, `AllMembers` ou `SpecificMembers`).",
            "nullable": true
          },
          "allowedUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Lista de utilizadores com acesso quando a visibilidade é `SpecificMembers`.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para criar uma nova lista."
      },
...
      "ListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador do item.",
            "format": "uuid"
          },
          "listId": {
            "type": "string",
            "description": "Identificador da lista.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Nome do item.",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Indica se o item está concluído."
          },
          "categoryId": {
            "type": "string",
            "description": "Identificador da categoria do item (opcional).",
            "format": "uuid",
            "nullable": true
          },
          "assigneeUserId": {
            "type": "string",
            "description": "Identificador do utilizador a quem o item estÃ¡ atribuÃ­do (opcional).",
            "format": "uuid",
            "nullable": true
          },
          "note": {
            "type": "string",
            "description": "Nota opcional associada ao item.",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "description": "URL (ou data URL) de uma foto associada ao item (opcional).",
            "nullable": true
          },
          "completedAtUtc": {
            "type": "string",
            "description": "Data/hora de conclusão (UTC), se concluído.",
            "format": "date-time",
            "nullable": true
          },
          "completedByUserId": {
            "type": "string",
            "description": "Identificador do utilizador que concluiu o item, se aplicável.",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Resposta com informação de um item de lista."
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador da lista.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Nome da lista.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Tipo de lista (`Shopping`, `Tasks` ou `Custom`).",
            "nullable": true
          },
          "colorHex": {
            "type": "string",
            "description": "Cor da lista em hexadecimal.",
            "nullable": true
          },
          "coverImageUrl": {
            "type": "string",
            "description": "URL da imagem de capa da lista.",
            "nullable": true
          },
          "ownerUserId": {
            "type": "string",
            "description": "Identificador do utilizador proprietário (criador).",
            "format": "uuid"
          },
          "visibilityMode": {
            "type": "string",
            "description": "Modo de visibilidade da lista (`Private`, `AllMembers` ou `SpecificMembers`).",
            "nullable": true
          },
          "allowedUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Lista de utilizadores com acesso quando a visibilidade é `SpecificMembers`.",
            "nullable": true
          },
          "sharedWithMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedListMemberPreviewResponse"
            },
            "description": "Membros com os quais a lista é partilhada (ou seja, quem tem acesso à lista).",
            "nullable": true
          },
          "itemsCount": {
            "type": "integer",
            "description": "Número total de itens na lista.",
            "format": "int32"
          },
          "completedCount": {
            "type": "integer",
            "description": "Número de itens concluídos.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Resposta com informação de uma lista."
      },
...
      "SharedListMemberPreviewResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Identificador do utilizador.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Nome do utilizador.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Resposta com informação mínima de um membro (para pré-visualizações, como avatares)."
      },
...
      "UpdateListItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Novo nome do item (opcional).",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Indica se o item passa a estar concluído (opcional).",
            "nullable": true
          },
          "categoryId": {
            "description": "Identificador da categoria do item (opcional).",
            "nullable": true
          },
          "assigneeUserId": {
            "description": "Identificador do utilizador a quem o item estÃ¡ atribuÃ­do (opcional).",
            "nullable": true
          },
          "note": {
            "description": "Nota do item (opcional).",
            "nullable": true
          },
          "photoUrl": {
            "description": "URL (ou data URL) da foto do item (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para atualizar um item de lista."
      },
      "UpdateListRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da lista.",
            "nullable": true
          },
          "colorHex": {
            "type": "string",
            "description": "Cor da lista em hexadecimal (ex.: `#FFAA00`).",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Tipo de lista (`Shopping`, `Tasks` ou `Custom`).",
            "nullable": true
          },
          "visibilityMode": {
            "type": "string",
            "description": "Novo modo de visibilidade (opcional).",
            "nullable": true
          },
          "allowedUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Lista de utilizadores com acesso, quando aplicável (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para atualizar uma lista."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Insere o token JWT assim: **Bearer {token}**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}